tools/python/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
plexapi==4.17.2
python-dotenv==1.0.0
requests==2.31.0
numpy>=1.26.0
scikit-learn>=1.4.0
pandas>=2.2.0
